@@ -1,4 +1,4 @@
-// swift-tools-version: 5.7
+// swift-tools-version: 5.8
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -24,6 +24,9 @@ let package = Package(
             name: "MoyaAPIClient",
             dependencies: [
                 .product(name: "Moya", package: "Moya"),
+            ],
+            swiftSettings: [
+                .unsafeFlags(["-Xfrontend", "-strict-concurrency=complete"]),
             ]
         ),
         .testTarget(
